Stop explicitly building extensions
