+++ msgraph/cli/core/installed_extensions.py
@@ -1,24 +1,25 @@
-installed_extensions = [ 'azext_applications',
- 'azext_calendar',
- 'azext_changenotifications',
- 'azext_cloudcommunications',
- 'azext_crossdeviceexperiences',
- 'azext_directoryobjects',
- 'azext_education',
- 'azext_files',
- 'azext_identitydirectorymanagement',
- 'azext_identitysignins',
- 'azext_mail',
- 'azext_notes',
- 'azext_people',
- 'azext_personalcontacts',
- 'azext_planner',
- 'azext_reports',
- 'azext_schemaextensions',
- 'azext_security',
- 'azext_sites',
- 'azext_teams',
- 'azext_users',
- 'azext_usersactions',
- 'azext_usersfunctions',
+installed_extensions = [
+    'azext_schemaextensions',
+    'azext_cloudcommunications',
+    'azext_calendar',
+    'azext_usersfunctions',
+    'azext_security',
+    'azext_mail',
+    'azext_planner',
+    'azext_education',
+    'azext_sites',
+    'azext_changenotifications',
+    'azext_notes',
+    'azext_users',
+    'azext_usersactions',
+    'azext_people',
+    'azext_teams',
+    'azext_applications',
+    'azext_files',
+    'azext_identitydirectorymanagement',
+    'azext_directoryobjects',
+    'azext_crossdeviceexperiences',
+    'azext_personalcontacts',
+    'azext_identitysignins',
+    'azext_reports',
 ]
